refactor(ai): move WorkflowEngine to ai/drafts/ (issue #2760, Option C)
WorkflowEngine and workflow_definitions are complete scaffolding but have
never been wired to a GUI entry point.  Moving them to ai/drafts/ preserves
~1300 lines of work while making it clear they are not part of the public
API surface.  The drafts/__init__.py explains the intent and cross-references
the issue number.

- git mv workflow_engine.py + workflow_definitions.py -> ai/drafts/
- ai/drafts/__init__.py added with status comment (issue #2760)
- workflow_definitions.py: update import to drafts.workflow_engine
- ai/__init__.py: remove all workflow exports from TYPE_CHECKING block,
  _LAZY dict, and __all__; update module docstring
- scripts/monolith_baseline.txt: update file paths

Closes #2760
